[x/programs] add program abstraction (#660)
* Move fixtures into tests directory


* Abstract engine from runtime


* React to changes


* Nits


* Review comments


* Clarify debug mode


* Nit


* Add host abstraction


* React to changes


* Cleanup


* Move program from runtime


* Add program abstraction


* Implement runtime instance and react to changes


* Nit


* Add RegisterImportWrapFn


* Update examples


* Show example of wrap import function impl


* Remove stale docs


* Impliment pstate with wrap functionality


* license header lint

* remove hardcoded wrapper funcitons, fixed pstate bugs

* lint

* lint

* Revert "Merge branch 'dynamic_args' into refactor-statekey"

This reverts commit 5c359d986482310aa38c71322a04ab84c95bb2fb, reversing
changes made to cee952cf7ae6f97567a452456da7770d0d232726.

* Revert "Revert "Merge branch 'dynamic_args' into refactor-statekey""

This reverts commit 15b84b5fb0313a583ff8eaf5e91b116ceb0aff95.

* lint

* memory tests

* create WAT program instead of memory.wasm

---------

+++ x/programs/rust/wasmlanche_sdk/src/host/state.rs
@@ -8,7 +8,7 @@ use borsh::{to_vec, BorshSerialize};
 #[link(wasm_import_module = "state")]
 extern "C" {
     #[link_name = "put"]
-    fn _put(caller: i64, key: i64, value: i64) -> i32;
+    fn _put(caller: i64, key: i64, value: i64) -> i64;
 
     #[link_name = "get"]
     fn _get(caller: i64, key: i64) -> i64;
